Add features & Clean up code

+++ src/main/java/org/az/climator/resource/ProfileUpdateResource.java
@@ -2,22 +2,26 @@ package org.az.climator.resource;
 
 import io.quarkus.elytron.security.common.BcryptUtil;
 import org.az.climator.entity.UserEntity;
-import org.az.climator.validation.PasswordValidation.ValidPassword;
-import org.eclipse.microprofile.openapi.annotations.enums.SecuritySchemeType;
+import org.az.climator.services.JWTService;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
-import org.eclipse.microprofile.openapi.annotations.security.SecurityScheme;
 
 import javax.annotation.security.RolesAllowed;
+import javax.inject.Inject;
 import javax.transaction.Transactional;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
 
 @Path("/manage")
 public class ProfileUpdateResource {
 
-//    TODO: Add security features: JWT
+    @Inject
+    JWTService jwtService;
+
+    @Context
+    SecurityContext securityContext;
 
     @PUT
     @Transactional
@@ -27,13 +31,12 @@ public class ProfileUpdateResource {
         responseCode = "200",
         description = "Update User's Password"
     )
-    public Response updatePasswordById(@PathParam("id") Long id, @ValidPassword @QueryParam("newPass") String newPass) {
-        UserEntity user = UserEntity.findById(id);
-        String prevPassword = user.encodedPassword;
-        user.encodedPassword = BcryptUtil.bcryptHash(newPass);
+    public Response updatePasswordById(@PathParam("id") Long id, @QueryParam("newPass") String newPass) {
 
-        return user.encodedPassword != prevPassword ? Response.ok().build()
-                :  Response.status(Response.Status.BAD_REQUEST).build();
+        UserEntity user = UserEntity.findById(id);
+        user.encodedPassword = BcryptUtil.bcryptHash(newPass);
+        return jwtService.verifyJWT(id,securityContext) ? Response.ok().build() :
+                Response.status(Response.Status.BAD_REQUEST).build();
     }
 
     @DELETE
@@ -44,17 +47,12 @@ public class ProfileUpdateResource {
             responseCode = "204",
             description = "Delete User"
     )
+    @Produces(MediaType.TEXT_PLAIN)
     public Response deleteById(@PathParam("id") Long id) {
-        boolean deleted = UserEntity.deleteById(id);
-        return deleted ? Response.noContent().build() :
-                Response.status(Response.Status.BAD_REQUEST).build();
+        if (jwtService.verifyJWT(id, securityContext)){
+            UserEntity.deleteById(id);
+            return Response.noContent().build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
-
-    @GET
-    @RolesAllowed("user")
-    @Path("/me")
-    public String me(@Context SecurityContext securityContext) {
-        return securityContext.getUserPrincipal().getName();
-    }
-
 }
